pybackend/requirements.txt
fastapi
uvicorn
pymongo
facenet-pytorch
torch
torchvision
numpy
Pillow
opencv-python-headless
